Split cloud auth test suites
